Moves editable component title to CodapComponent
