+++ v3/src/components/data-summary.scss
@@ -12,7 +12,7 @@
   .summary-inspector-drop {
     position: absolute;
     left: 10px;
-    bottom: 10px;
+    bottom: 35px;
     border: 1px solid gray;
     padding: 5px 10px;
     background: #eee;
@@ -27,14 +27,14 @@
   .summary-attribute-info {
     position: absolute;
     left: 160px;
-    bottom: 10px;
+    bottom: 35px;
     padding: 5px;
   }
 
   .profiler-button {
     position: absolute;
     right: 10px;
-    bottom: 10px;
+    bottom: 35px;
     padding: 5px 10px;
     background: #eee;
   }
