+++ v3/src/components/graph/components/graph-component.tsx
@@ -1,12 +1,11 @@
 import {Button} from '@chakra-ui/react'
 import {observer} from "mobx-react-lite"
-import React, {useEffect, useRef, useState} from "react"
+import React, {useEffect, useRef} from "react"
 import {useResizeDetector} from "react-resize-detector"
 import {useMemo} from "use-memo-one"
 import {DataBroker} from "../../../data-model/data-broker"
 import {DataSetContext} from "../../../hooks/use-data-set-context"
 import {InstanceIdContext, useNextInstanceId} from "../../../hooks/use-instance-id-context"
-import {EditableComponentTitle} from '../../editable-component-title'
 import {MovableLineModel, MovableValueModel} from "../adornments/adornment-models"
 import {NumericAxisModel} from "../models/axis-model"
 import {GraphLayout, GraphLayoutContext} from "../models/graph-layout"
@@ -32,38 +31,29 @@ export const GraphComponent = observer(({broker}: IProps) => {
   const layout = useMemo(() => new GraphLayout(), [])
   const {width, height, ref: graphRef} = useResizeDetector({refreshMode: "debounce", refreshRate: 200})
   const animationIsOn = useRef(true)
-  const [componentTitle, setComponentTitle] = useState("")
-
-  const handleTitleChange = (title?: string) => {
-    title && setComponentTitle(title)
-  }
 
   useEffect(() => {
     (width != null) && (height != null) && layout.setGraphExtent(width, height)
   }, [width, height, layout])
 
   return (
-    <>
-      <EditableComponentTitle componentTitle={componentTitle}
-          onEndEdit={handleTitleChange} />
-      <DataSetContext.Provider value={broker?.last}>
-        <InstanceIdContext.Provider value={instanceId}>
-          <GraphLayoutContext.Provider value={layout}>
-            <Graph model={defaultGraphModel} graphRef={graphRef} animationIsOn={animationIsOn}/>
-            <Button
-              className='graph-plot-choice'
-              size="xs"
-              onClick={() => {
-                const currPlotType = defaultGraphModel.plotType
-                animationIsOn.current = true
-                defaultGraphModel.setPlotType(currPlotType === 'scatterPlot' ? 'dotPlot' : 'scatterPlot')
-              }}
-            >
-              {defaultGraphModel.plotType === 'scatterPlot' ? 'Dot' : 'Scatter'} Plot
-            </Button>
-          </GraphLayoutContext.Provider>
-        </InstanceIdContext.Provider>
-      </DataSetContext.Provider>
-    </>
+    <DataSetContext.Provider value={broker?.last}>
+      <InstanceIdContext.Provider value={instanceId}>
+        <GraphLayoutContext.Provider value={layout}>
+          <Graph model={defaultGraphModel} graphRef={graphRef} animationIsOn={animationIsOn}/>
+          <Button
+            className='graph-plot-choice'
+            size="xs"
+            onClick={() => {
+              const currPlotType = defaultGraphModel.plotType
+              animationIsOn.current = true
+              defaultGraphModel.setPlotType(currPlotType === 'scatterPlot' ? 'dotPlot' : 'scatterPlot')
+            }}
+          >
+            {defaultGraphModel.plotType === 'scatterPlot' ? 'Dot' : 'Scatter'} Plot
+          </Button>
+        </GraphLayoutContext.Provider>
+      </InstanceIdContext.Provider>
+    </DataSetContext.Provider>
   )
 })
